0.1
A string variable generated procedurally by a simple algorithm, which i myself programmed. The variable represents a door with, for now, walls and a door. The door is also procedurally positioned. This could also be done in C but i could not be arsed to find the substitutes for std library functions.